+++ CHANGELOG.md
@@ -1,8 +1,9 @@
 # Change Log
 
 ## [unreleased](https://github.com/deltaDNA/ios-smartads-sdk/releases/tag/) ()
-### Added 
+### Added
 - Support for Supersonic (IronSource)
+- Support for Facebook Audience Network
 
 ## [1.3.0](https://github.com/deltaDNA/ios-smartads-sdk/releases/tag/1.3.0) (2017-03-13)
 ### Added